pyproject.toml
[project]
name = "2img-to-gif"
version = "0.0.1"
description = "This Python program produces a GIF that seamlessly transforms from the first to the second image when two different images are passed."
authors = [
    { name = "shinshin86", email = "shinshin86npm@gmail.com" }
]
dependencies = [
    "gradio>=3.47.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
